fix: HAProxy channel-active and user-event triggers for both `AUTO` and `ON` mode
