+++ haproxy/src/main/kotlin/org/jire/netty/haproxy/HAProxyMessageHandler.kt
@@ -9,6 +9,7 @@ import io.netty.channel.SimpleChannelInboundHandler
 import io.netty.handler.codec.haproxy.HAProxyCommand
 import io.netty.handler.codec.haproxy.HAProxyMessage
 import io.netty.handler.codec.haproxy.HAProxyMessageDecoder
+import io.netty.handler.timeout.IdleStateEvent
 import org.jire.netty.haproxy.HAProxyAttributes.haproxyAttribute
 import org.jire.netty.haproxy.HAProxyHandlerNames.HAPROXY_CHANNEL_INITIALIZER_NAME
 import org.jire.netty.haproxy.HAProxyHandlerNames.HAPROXY_IDLE_STATE_HANDLER_NAME
@@ -25,6 +26,27 @@ import org.jire.netty.haproxy.HAProxyHandlerNames.HAPROXY_IDLE_STATE_HANDLER_NAM
 public class HAProxyMessageHandler<C : Channel>(
     private val childInitializer: ChannelInitializer<C>,
 ) : SimpleChannelInboundHandler<HAProxyMessage>(true) {
+    override fun channelActive(ctx: ChannelHandlerContext) {
+        // Because auto-read may be disabled, we need to trigger the handling
+        ctx.read()
+
+        ctx.fireChannelActive()
+    }
+
+    override fun userEventTriggered(
+        ctx: ChannelHandlerContext,
+        evt: Any,
+    ) {
+        ctx.fireUserEventTriggered(evt)
+
+        if (evt is IdleStateEvent) {
+            logger.debug {
+                "Channel has gone idle during HAProxy message handling, closing for ${ctx.channel()}"
+            }
+            ctx.close()
+        }
+    }
+
     override fun channelRead0(
         ctx: ChannelHandlerContext,
         msg: HAProxyMessage,
